+++ src/infra/clients/ImageStorageClient.ts
@@ -1,15 +1,16 @@
 import { AxiosResponse } from 'axios';
 import { FileObject } from 'src/types/File';
 import { ImageStorageClientContract } from 'src/types/clients/ImageStorageClient';
+import Configs from 'src/types/shared/configs';
 
 export default class ImageStorageClient {
     private readonly _logger;
     private readonly _configs;
     private readonly _httpRequest;
 
-    constructor({ logger, configs, httpRequest }: ImageStorageClientContract) {
+    constructor({ logger, httpRequest }: ImageStorageClientContract) {
         this._logger = logger;
-        this._configs = configs;
+        this._configs = require('../../../../tablerise.environment.js') as Configs;
         this._httpRequest = httpRequest;
     }
 
@@ -17,7 +18,7 @@ export default class ImageStorageClient {
         this._logger('info', 'Upload - ImageStorageClient');
         const { baseUrl, authorization, endpoints } = this._configs.api.imgur;
 
-        const url = baseUrl + endpoints.postImage;
+        const url = `${baseUrl + endpoints.postImage}` ;
 
         const imageUploaded =
             process.env.NODE_ENV === 'production'
