+++ src/types/modules/domains/common/helpers/parametersWrapper.d.ts
@@ -1,0 +1,16 @@
+import { routeInstance } from '@tablerise/auto-swagger';
+
+export interface RouteDeclareParams {
+    name: string;
+    location: string;
+    required: boolean;
+    type: string;
+}
+
+export interface ParamName {
+    name: string;
+    type: string;
+    required?: 'off';
+}
+
+export type RouteWrapperDeclared = routeInstance;
